requirements.txt
Django==5.2.3
djangorestframework==3.16.0
drf_spectacular==0.28.0
python-decouple==3.8
